Improve Stripe manifest: extract to separate file, enhance schemas and templates
DX/UX improvements from review feedback:

- Extract Manifest() to dedicated manifest.go (follows hubspot/amadeus/square
  pattern — keeps stripe.go focused on struct, auth, and HTTP lifecycle)
- Add additionalProperties: false to all parameter schemas to prevent
  unexpected fields and improve validation error messages
- Add format hints (email, uri) and minimum/minItems constraints
- Improve descriptions with examples (e.g. "1050 cents = $10.50"),
  E.164 phone format, ISO 4217 currency codes
- Add risk-tiered templates following square/amadeus pattern:
  - Two subscription list templates: active-only vs any status
  - Two refund templates: capped at $99.99 vs any amount (trusted agents)
- Organize templates by risk level (read-only → low → medium → high)
  with section comments for clarity
- Strengthen refund warning: "WARNING: this moves real money and cannot
  be undone"
