+++ connectors/stripe/stripe.go
@@ -103,279 +103,6 @@ func newForTest(client *http.Client, baseURL string) *StripeConnector {
 // ID returns "stripe", matching the connectors.id in the database.
 func (c *StripeConnector) ID() string { return "stripe" }
 
-// Manifest returns the connector's metadata manifest. Used by the server to
-// auto-seed DB rows on startup, replacing manual seed.go files.
-func (c *StripeConnector) Manifest() *connectors.ConnectorManifest {
-	return &connectors.ConnectorManifest{
-		ID:          "stripe",
-		Name:        "Stripe",
-		Description: "Stripe integration for payments, invoicing, and billing",
-		Actions: []connectors.ManifestAction{
-			{
-				ActionType:  "stripe.create_customer",
-				Name:        "Create Customer",
-				Description: "Create a new customer record — foundational for all other Stripe operations",
-				RiskLevel:   "low",
-				ParametersSchema: json.RawMessage(connectors.TrimIndent(`{
-					"type": "object",
-					"required": ["email"],
-					"properties": {
-						"email": {
-							"type": "string",
-							"description": "Customer email address"
-						},
-						"name": {
-							"type": "string",
-							"description": "Customer full name"
-						},
-						"description": {
-							"type": "string",
-							"description": "Free-form description of the customer"
-						},
-						"phone": {
-							"type": "string",
-							"description": "Customer phone number"
-						},
-						"metadata": {
-							"type": "object",
-							"description": "Key-value pairs for storing additional information (max 50 keys)",
-							"additionalProperties": { "type": "string" }
-						}
-					}
-				}`)),
-			},
-			{
-				ActionType:  "stripe.create_invoice",
-				Name:        "Create Invoice",
-				Description: "Create and optionally send an invoice with line items",
-				RiskLevel:   "medium",
-				ParametersSchema: json.RawMessage(connectors.TrimIndent(`{
-					"type": "object",
-					"required": ["customer_id"],
-					"properties": {
-						"customer_id": {
-							"type": "string",
-							"description": "Stripe customer ID (cus_...)"
-						},
-						"description": {
-							"type": "string",
-							"description": "Invoice description"
-						},
-						"due_date": {
-							"type": "integer",
-							"description": "Due date as Unix timestamp"
-						},
-						"auto_advance": {
-							"type": "boolean",
-							"default": true,
-							"description": "Automatically finalize and send the invoice"
-						},
-						"currency": {
-							"type": "string",
-							"default": "usd",
-							"description": "Three-letter ISO currency code (defaults to usd)"
-						},
-						"line_items": {
-							"type": "array",
-							"description": "Invoice line items",
-							"items": {
-								"type": "object",
-								"properties": {
-									"description": {
-										"type": "string",
-										"description": "Line item description"
-									},
-									"amount": {
-										"type": "integer",
-										"description": "Amount in cents (must be positive)"
-									},
-									"quantity": {
-										"type": "integer",
-										"description": "Quantity (defaults to 1)"
-									}
-								}
-							}
-						},
-						"metadata": {
-							"type": "object",
-							"description": "Key-value pairs for storing additional information (max 50 keys)",
-							"additionalProperties": { "type": "string" }
-						}
-					}
-				}`)),
-			},
-			{
-				ActionType:  "stripe.issue_refund",
-				Name:        "Issue Refund",
-				Description: "Refund a charge or payment intent — high risk: moves real money",
-				RiskLevel:   "high",
-				ParametersSchema: json.RawMessage(connectors.TrimIndent(`{
-					"type": "object",
-					"properties": {
-						"payment_intent_id": {
-							"type": "string",
-							"description": "Payment intent ID (pi_...) — provide this or charge_id"
-						},
-						"charge_id": {
-							"type": "string",
-							"description": "Charge ID (ch_...) — provide this or payment_intent_id"
-						},
-						"amount": {
-							"type": "integer",
-							"description": "Refund amount in cents (omit for full refund)"
-						},
-						"reason": {
-							"type": "string",
-							"enum": ["duplicate", "fraudulent", "requested_by_customer"],
-							"description": "Reason for the refund"
-						},
-						"metadata": {
-							"type": "object",
-							"description": "Key-value pairs for storing additional information (max 50 keys)",
-							"additionalProperties": { "type": "string" }
-						}
-					}
-				}`)),
-			},
-			{
-				ActionType:  "stripe.list_subscriptions",
-				Name:        "List Subscriptions",
-				Description: "List subscriptions, optionally filtered by customer or status",
-				RiskLevel:   "low",
-				ParametersSchema: json.RawMessage(connectors.TrimIndent(`{
-					"type": "object",
-					"properties": {
-						"customer_id": {
-							"type": "string",
-							"description": "Filter by Stripe customer ID (cus_...)"
-						},
-						"status": {
-							"type": "string",
-							"enum": ["active", "past_due", "canceled", "unpaid", "trialing", "all"],
-							"description": "Filter by subscription status"
-						},
-						"price_id": {
-							"type": "string",
-							"description": "Filter by price ID (price_...)"
-						},
-						"limit": {
-							"type": "integer",
-							"default": 10,
-							"minimum": 1,
-							"maximum": 100,
-							"description": "Number of subscriptions to return (default 10, max 100)"
-						}
-					}
-				}`)),
-			},
-			{
-				ActionType:  "stripe.create_payment_link",
-				Name:        "Create Payment Link",
-				Description: "Create a shareable payment link for one-time or recurring purchases",
-				RiskLevel:   "medium",
-				ParametersSchema: json.RawMessage(connectors.TrimIndent(`{
-					"type": "object",
-					"required": ["line_items"],
-					"properties": {
-						"line_items": {
-							"type": "array",
-							"description": "Products to include in the payment link",
-							"items": {
-								"type": "object",
-								"required": ["price_id", "quantity"],
-								"properties": {
-									"price_id": {
-										"type": "string",
-										"description": "Stripe price ID (price_...)"
-									},
-									"quantity": {
-										"type": "integer",
-										"description": "Quantity of the product"
-									}
-								}
-							}
-						},
-						"after_completion": {
-							"type": "string",
-							"description": "Redirect URL after successful payment"
-						},
-						"allow_promotion_codes": {
-							"type": "boolean",
-							"description": "Allow customers to enter promotion codes"
-						},
-						"metadata": {
-							"type": "object",
-							"description": "Key-value pairs for storing additional information (max 50 keys)",
-							"additionalProperties": { "type": "string" }
-						}
-					}
-				}`)),
-			},
-			{
-				ActionType:  "stripe.get_balance",
-				Name:        "Get Balance",
-				Description: "Retrieve the current account balance — useful for monitoring cash flow",
-				RiskLevel:   "low",
-				ParametersSchema: json.RawMessage(connectors.TrimIndent(`{
-					"type": "object",
-					"properties": {}
-				}`)),
-			},
-		},
-		RequiredCredentials: []connectors.ManifestCredential{
-			{
-				Service:         "stripe",
-				AuthType:        "api_key",
-				InstructionsURL: "https://docs.stripe.com/keys",
-			},
-		},
-		Templates: []connectors.ManifestTemplate{
-			{
-				ID:          "tpl_stripe_create_customers",
-				ActionType:  "stripe.create_customer",
-				Name:        "Create customers",
-				Description: "Agent can create new customer records with any details.",
-				Parameters:  json.RawMessage(`{"email":"*","name":"*","description":"*","phone":"*"}`),
-			},
-			{
-				ID:          "tpl_stripe_create_invoices",
-				ActionType:  "stripe.create_invoice",
-				Name:        "Create invoices",
-				Description: "Agent can create and send invoices for any customer.",
-				Parameters:  json.RawMessage(`{"customer_id":"*","description":"*","line_items":"*"}`),
-			},
-			{
-				ID:          "tpl_stripe_issue_refund_capped",
-				ActionType:  "stripe.issue_refund",
-				Name:        "Issue refunds up to $99.99",
-				Description: "Agent can issue refunds up to 9999 cents ($99.99). Amount is constrained by pattern to prevent large refunds.",
-				Parameters:  json.RawMessage(`{"payment_intent_id":"*","charge_id":"*","amount":{"$pattern":"^[1-9]\\d{0,3}$"},"reason":"*"}`),
-			},
-			{
-				ID:          "tpl_stripe_list_subscriptions",
-				ActionType:  "stripe.list_subscriptions",
-				Name:        "List active subscriptions",
-				Description: "Agent can list active subscriptions for any customer.",
-				Parameters:  json.RawMessage(`{"customer_id":"*","status":"active","limit":"*"}`),
-			},
-			{
-				ID:          "tpl_stripe_create_payment_links",
-				ActionType:  "stripe.create_payment_link",
-				Name:        "Create payment links",
-				Description: "Agent can create shareable payment links for any products.",
-				Parameters:  json.RawMessage(`{"line_items":"*","after_completion":"*","allow_promotion_codes":"*"}`),
-			},
-			{
-				ID:          "tpl_stripe_get_balance",
-				ActionType:  "stripe.get_balance",
-				Name:        "Check account balance",
-				Description: "Agent can retrieve the current Stripe account balance.",
-				Parameters:  json.RawMessage(`{}`),
-			},
-		},
-	}
-}
-
 // Actions returns the registered action handlers keyed by action_type.
 func (c *StripeConnector) Actions() map[string]connectors.Action {
 	return map[string]connectors.Action{
